+++ components/ColorBlock.tsx
@@ -11,14 +11,14 @@ export default function (props: iColorBlock) {
   const { color, ...p } = props;
   return (
     <div
-      class="color-block outline-cobalto m-0.5"
+      class="color-block m-0.5"
       tabIndex={color === null ? -1 : 0}
       {...p}
     >
       <div
-        class={`aspect-square rounded bg-${color ? color : "blanco"} ${
-          color === "blanco" ? "border-1 border-hierro" : null
-        }`}
+        class={`aspect-square rounded ${
+          color ? `bg-${color}` : "clr-bg-primary"
+        } ${color === "blanco" || color === "obsidiana" ? "border-1" : null}`}
         title={color ? color : ""}
       />
     </div>