+++ fresh.gen.ts
@@ -8,7 +8,8 @@ import * as $1 from "./routes/api/joke.ts";
 import * as $2 from "./routes/atoms/color.tsx";
 import * as $3 from "./routes/atoms/input.tsx";
 import * as $4 from "./routes/atoms/text.tsx";
-import * as $5 from "./routes/index.tsx";
+import * as $5 from "./routes/atoms/theme.tsx";
+import * as $6 from "./routes/index.tsx";
 import * as $$0 from "./islands/ColorDisplay.tsx";
 import * as $$1 from "./islands/Navigation.tsx";
 
@@ -19,7 +20,8 @@ const manifest = {
     "./routes/atoms/color.tsx": $2,
     "./routes/atoms/input.tsx": $3,
     "./routes/atoms/text.tsx": $4,
-    "./routes/index.tsx": $5,
+    "./routes/atoms/theme.tsx": $5,
+    "./routes/index.tsx": $6,
   },
   islands: {
     "./islands/ColorDisplay.tsx": $$0,