+++ routes/_app.tsx
@@ -8,7 +8,7 @@ export default function App({ Component }: AppProps) {
         <title>Ana Components Docs</title>
         <link
           rel="stylesheet"
-          href="https://cdn.jsdelivr.net/gh/CarcajadaArtificial/ana-components@v0.0.13/static/styles.css"
+          href="https://cdn.jsdelivr.net/gh/CarcajadaArtificial/ana-components@v0.0.20/static/styles.css"
         />
       </Head>
       <Component />